--- ClariseMain/src/main/java/com/clairvoyant/clarise/configuration/WebConfig.java
@@ -1,19 +1,0 @@
-//package com.clairvoyant.clarise.configuration;
-//
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.web.servlet.config.annotation.CorsRegistry;
-//import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-//
-//
-//@Configuration
-//public class WebConfig implements WebMvcConfigurer {
-//
-//    @Override
-//    public void addCorsMappings(CorsRegistry registry) {
-//        registry.addMapping("/**")
-//                .allowedOrigins("*")
-//                .allowedMethods("HEAD", "GET", "PUT", "POST", "DELETE", "PATCH")
-//                .allowedHeaders("*")
-//                .allowCredentials(true);
-//    }
-//}
