+++ ClariseMain/src/main/java/com/clairvoyant/clarise/graphql/resolver/mutation/CategoryMutationResolver.java
@@ -1,5 +1,5 @@
 
-package com.clairvoyant.clarise.graphql;
+package com.clairvoyant.clarise.graphql.resolver.mutation;
 
 
 import com.clairvoyant.clarise.enums.Status;
