+++ ClariseMain/src/main/java/com/clairvoyant/clarise/graphql/resolver/mutation/SkillCategoryMutationResolver.java
@@ -1,4 +1,4 @@
-package com.clairvoyant.clarise.graphql;
+package com.clairvoyant.clarise.graphql.resolver.mutation;
 
 import com.clairvoyant.clarise.dto.SkillCategoryDto;
 import com.clairvoyant.clarise.model.Category;
