+++ ClariseMain/src/test/java/com/clairvoyant/clarise/SkillBaseMainApplicationTests.java
@@ -1,9 +1,7 @@
 package com.clairvoyant.clarise;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-//@SpringBootTest
 class SkillBaseMainApplicationTests {
 
 	@Test
